Refactor and enhance code quality across multiple modules
- Improved docstrings for clarity in BaseMutation, BaseCrossover, and BaseSelection classes.
- Refactored code formatting for better readability in PolynomialMutation.
- Cleaned up whitespace inconsistencies in test files for binary and real evaluators.
- Enhanced assertions and formatting in test cases for better clarity and consistency.
- Updated test_gaussian.py to remove unnecessary imports and improve assertion clarity.
- Refactored test_mutation_inner_methods.py for improved assertion formatting.
- Enhanced test_ablation_decorators.py with clearer docstrings and improved test structure.
- Fixed formatting issues in test_fixes.py for better readability and consistency.

+++ src/malthusjax/operators/mutation/real.py
@@ -270,9 +270,9 @@ class PolynomialMutation(BaseMutation[RealGenome, RealGenomeConfig, RealPopulati
         half = jnp.array(0.5, dtype=dtype)
         two = jnp.array(2.0, dtype=dtype)
         exponent = one / (eta + one)
-        mask_val = jax.random.bernoulli(
-            k_mask, p=self.mutation_rate, shape=config.shape
-        ).astype(dtype)
+        mask_val = jax.random.bernoulli(k_mask, p=self.mutation_rate, shape=config.shape).astype(
+            dtype
+        )
         u = jax.random.uniform(k_val, shape=config.shape, dtype=dtype)
         delta_q = jnp.where(
             u <= half,
